account props try to work around no wallet event
