+++ 5-DAO/dao/client/src/pages/Marketplace.jsx
@@ -1,6 +1,28 @@
+import { useEffect, useState } from 'react';
+import Sdk from '@radixdlt/alphanet-walletextension-sdk';
+
 import BuyTokens from '../components/BuyTokens';
 
 const Marketplace = () => {
+  const [account, setAccount] = useState(
+    'account_tdx_a_1qd9eafyqjh750uv7scsy474xdceh2x2cjqdccus5k0ls06kddh'
+  );
+
+  const sdk = Sdk();
+
+  useEffect(() => {
+    const getAddress = async () => {
+      const result = await sdk.request({
+        accountAddresses: {},
+      });
+      console.log('accountAddresses: ', result.value);
+      const { accountAddresses } = result.value;
+      setAccount(accountAddresses[0].address);
+    };
+    getAddress();
+    return () => {};
+  }, [sdk]);
+
   return (
     <div>
       <h2 className="text-2xl font-bold mb-2">
@@ -10,7 +32,7 @@ const Marketplace = () => {
         Find new projects to support, contribute to, sell or trade member tokens
         with other community members.
       </p>
-      <BuyTokens />
+      <BuyTokens account={account} />
     </div>
   );
 };
